refactor: InputHandler 및 OutputHandler에 입/출력 로직 캡슐화
InputHandler 및 OutputHandler에 입/출력 로직 캡슐화

- InputHandler: 입력 데이터 처리를 담당하는 클래스
- OutputHandler: 출력 데이터 처리를 담당하는 클래스
